NAS-135059 / 25.10 / Flag smbhash of '*' as broken for passdb insertion (#16137)
It appears that some users may have a wildcard character with
accounts that have SMB enabled. This commit shifts NT hash
conversion from hex value to earlier and raises a ValueError
if it fails. The ValueError will get picked up when doing bulk
sync of passdb file and converted into an alert for user to
redo their password.