Convert ID_TYPE_BOTH to either USER or GROUP
This commit changes how we convert ID_TYPE_BOTH into a string for
backend consumption. Originally we were translating it to 'BOTH'
with the understanding that the underlying SID maps to both a user
and a group in the NSS backend. This apparently got mapped to
'Unknown' in our UI, which appears very broken to our enterprise
userbase. This commit dumbs down the backend somewhat so that we
no longer report 'BOTH' and resolve to either USER or GROUP based
on the underlying sid_type of the winbind entry.