NAS-135051 / 25.10 / Handle edge case to make sure selected source is used to boot (#16129)
## Problem

If devices are attached to a virt instance at the time of creation, the
user can set the boot_priority of an attached disk to a value greater
than 1. This would override the intended boot_device that the user set
as the source_type.

## Solution

Ensure that the selected `source_type` has the highest boot_priority so
that the instance may boot from it even if devices are attached to it
during creation.