NAS-134751 / 25.10 / Allow subscribing to stats of all virt instances (#16192)
## Context
It was requested that instead of subscribing to only one instance to get
its metrics, collective metrics should be reported for virt instances so
UI does not has to subscribe to each instance separately. While at it,
an optimized endpoint has been added to retrieve the name of instances
as doing `virt.instance.query` would be expensive comparatively based
off the data we get there.