NAS-134963 / 25.10 / Better report validation error for 53 port when configuring virt (#16204)
## Problem

A user had adguard app configured where on wildcard IP the app was
consuming port 53. When he tried to configure virt, it errored out with
`Address already in use`.

## Solution

Make sure that we properly error out what is consuming port 53 so the
user has an idea of what actually is happening and what steps can be
taken to correct this.

i.e
```
[EINVAL] virt_global_update.pool: Port 53 is required for virtualization but is currently in use by the following services on wildcard IPs (0.0.0.0/::): 'adguard-home' application. Please reconfigure these services to bind to specific IP addresses instead of wildcard IPs.
```