NAS-134863 / 25.10 / Fix remote audit logging (#16219)
Remote audit logging required two small tweaks to get full function:
Removed the 'final' statement from the syslog audit log stanza and, to
avoid remote log filtering, changed the priority of the audit log
messages from `debug` to `info`.

Other tweaks:  fixed flake8 complaint, added a comment.

The majority of the changes are in CI where functional remote logging
tests are added. The remote logging test will run on HA systems only and
uses the standby node as the 'remote syslog server'.

The actual code changes are small, The majority of the changes are in
the testing modules.

NOTE: Unable to get stable test runs via jenkins. I was able to run the
test successfully (repeatedly) when the runner was outside of jenkins.