NAS-135306 / 25.10 / Only allow enterprise users to consume apps if they are licensed (#16251)
We were not checking if an enterprise user is licensed to use apps and
they could do so with consuming the API directly. Also while fixing this
i saw that some validation steps we perform should be done before we try
to mount docker related datasets which has been fixed/addressed as well.