Add ability to query utmp file to middleware
This commit adds ctype wrappers for utmp(5) related C APIs. These
APIs are generally not MT-SAFE and can be race-prone and so they
are protected behind a global threading lock. The utmp file is
exposed to middleware internally through the private API
system.security.sessions.query, but API schema is defined in such
a way that it can be easily placed in public API if required.

This commit is part of making middleware properly interact with
generic OS mechanisms for tracking and applying session count
limits for individual users as required by SRG-OS-000027-GPOS-00008.
There will be follow-up commits that ensure that middleware
login results in utmp entry creation and that middleware logout
results in clearing utmp entry. The follow-up commits will also
add explicit tests for the functionality.