NAS-135049 / 25.10 / Get rid of jobs in JSON-RPC 2.0 API (#16128)
Jobs are now an internal concept. From the most basic client’s point of
view, calling a job is equivalent to calling a standard method.

To switch to the new mode, include `{"legacy_jobs": false}` in your
initial `core.set_options` call. This method must be called immediately
after connecting to the API endpoint.

When a new job is initiated through a JSON-RPC 2.0 API call, its
`message_ids` field will contain the `id` of the call.
Therefore, when starting a new job, the client should listen for the
`added` event in the `core.get_jobs` subscription.
Additionally, the client should monitor `changed` events, as a `changed`
event with a new `message_ids` field value may be emitted if a method
call triggers a job that has already been scheduled.

Please refer to the "Jobs" page of the API documentation for the
example.