NAS-134717 / 25.10 / Allow specifying image OS for virt VMs (#16228)
### Context

RTC for incus VMs are by default set to UTC, which is fine for linux
base VMs but it is problematic for Windows VMs because Windows requires
RTC to be set to `localtime` in order to work correctly. To resolve
this, it is required that `image.os` property is set to `Windows`, by
doing this, incus will set rtc to localtime internally.

This PR introduces a `image_os` field which a user can specify at the
time of creating an instance or updating it. We specify a few popular OS
choices but let the user specify anything else he might want as well.