NAS-135489 / 25.10 / Do not enforce boot priority for ISO volumes (#16315)
## Context

We were enforcing that user specify boot priority when they added a disk
device which was based off an ISO volume so it was a conscious decision
on the user part in terms of knowing they want to boot from the ISO or
not. However with recent changes made to UI on how it handles boot
priority and allowing users to boot from X disk device, this is no
longer required and actually hurts usability hence this check is being
removed.