Fix passwd and shadow file generation
This commit fixes an error is passwd and shadow file generation
that was introduced by adding a synthetic account for the
container root user, the uid of this account is now explicitly
filtered out of the user query passed as render state to the
respective mako files. Test coverage is also added.