NAS-135550 / 25.10 / fix IndexError in account.update (#16341)
2 issues here being addressed:
1. on line 782 we have a check for `if not group`, when that branch is
true we add to a verrors object but never check for it. Since we don't
check for this verrors object, on line 784 we do not protect the
`group[0]` action and so we crash with an IndexError.
2. On line 650, we have the same exact logic, however, we raise a
`CallError` in that if branch which protects us from this IndexError
crash.

Fixes implemented:
1. protect the `group = group[0]` operation in an `else` branch to
prevent `IndexError`
2. be consistent and create a `verrors` object and add a
`verrors.check()` call before we move further down the logic