NAS-135551 / 25.10 / better validation error when group is None (#16343)
Related to PR https://github.com/truenas/middleware/pull/16341, testing
has shown the UI is sending `group: null` when someone tries to remove
the primary group from a user on update. This is an invalid
configuration and produces an error that isn't helpful. This changes the
logic in the update endpoint so that an informative validation error is
raised.