NAS-135475 / 25.10 / Do not allow to unset userns_idmap for builtins (#16361)
## Problem

We had validation in place to prevent changing `userns_idmap` attr for
builtin groups, however the validation was missing an edge case. We
checked if user tried to change the attr to something else but if the
attr was set and had `null ` in it - then it would bypass the existing
validation we have.

## Solution

Make sure we properly handle the case where someone could unset
`userns_idmap` for builtin users and also add integration tests to
properly validate this.