NAS-135308 / 25.04.1 / Fix virt instance attachment delegate (by Qubad786) (#16381)
## Problem

There are 2 problems here:

1. Missing attachment delegate for incus service
2. Mixed/convoluted logic for virt instance attachment delegate which
along with trying to account for (1) is also not handling zvols properly

Case (1) will be addressed in a subsequent PR, whereas for (2) we are
not accounting for zvols and also trying to mix logic in terms of what
should happen when delete is called.

## Solution

Fix the attachment delegate to properly account for zvols and also on
deletion make sure that we don't unset incus service but rather remove
that zvol from a virt instance for example if that zvol is being
deleted.

Original PR: https://github.com/truenas/middleware/pull/16379
Jira URL: https://ixsystems.atlassian.net/browse/NAS-135308

Co-authored-by: M. Rehan <mrehanlm93@gmail.com>