Relax security check for password aging data
If password authentication is disabled and user has never had a
password set, then we should not set the flag that a password
change is required. The reason for this is that the account may
still be used for an SSH session that is authenticated via key
and we need to allow normal login to succeed (and not prompt for
a password change).

(cherry picked from commit 27fbd61673a06315c64d119ce2695deb84c74e15)