NAS-135623 / 25.04.1 / Relax security check for password aging data (by anodos325) (#16384)
If password authentication is disabled and user has never had a password
set, then we should not set the flag that a password change is required.
The reason for this is that the account may still be used for an SSH
session that is authenticated via key and we need to allow normal login
to succeed (and not prompt for a password change).

Original PR: https://github.com/truenas/middleware/pull/16383
Jira URL: https://ixsystems.atlassian.net/browse/NAS-135623

Co-authored-by: Andrew Walker <awalker@ixsystems.com>