NAS-135663 / 25.04.2 / increase middlewared.service timeout (by yocalebo) (#16408)
We have a few users unable to upgrade to 25.04.0 because the
middlewared.service is being killed during boot process because our
default timeout is 4mins (240 seconds). I have a fear this is a
combination of pydantic and really poor performing boot drives. Either
way, we are adding more and more to the monolithic middlewared service
and so it's inevitable that service initialization takes longer.

Original PR: https://github.com/truenas/middleware/pull/16406
Jira URL: https://ixsystems.atlassian.net/browse/NAS-135663

Co-authored-by: Caleb <yocalebo@gmail.com>