NAS-135679 / 25.10 / NFS share allows host and network config that should be avoided. (#16416)
Using host entry `0.0.0.0` or network entry `0.0.0.0/0` to indicate
'allow everybody' results in unexpected and undesirable behavior.
The values are not explicitly excluded and some internet searches might
lead a user to use them to configure an NFS share for `everybody`.

This PR blocks those values with an error message suggesting a simple
alternative.

CI tests are added for these conditions.