Change how oplocks are handled for multiprotocol shares
This commit removes kernel oplocks in favor of disabling oplocks
on a per-share basis when they have been flagged for mixed-mode
use. This avoids issues observed in the field with kernel lease
breaks possibly blocking smbd main loop and causing client timeouts
as well allowing SMB leases globally. If client requests an SMB
lease on a multiprotocol share as part of an SMB2 create request
then no lease will be granted, but the open itself will succeed.
This matches older behavior in TrueNAS core and has not resulted
in any bug reports historically.