NAS-127189 / 25.10 / Add PAM-based session management for middleware (#16397)
This commit implements requirements for two specific STIGs:

SRG-OS-000021-GPOS-00005, SRG-OS-000027-GPOS-00008

**SRG-OS-000021-GPOS-00005** states that:
"The operating system must enforce the limit of three consecutive
invalid logon attempts by a user during a 15-minute time period."

Per discussions with the security team, this requirement covers all
authentication methods to the NAS. So with the same account one failed
auth attempt via API, a second via UI, and a third via SSH within the
same 15 minute period must result in the account being locked. This is
implemented through pam_faillock(8).

During scoping for this feature the security team agreed that this lock
may be a temporary 15 minute lock, with the account unlocking on
expiration.

**SRG-OS-000027-GPOS-00008**
"The operating system must limit the number of concurrent sessions to
ten for all accounts and/or account types."

Per discusions with the security team, this applies to interactive
sessions across administrative interfaces. This means a logged-in SSH
session should count against an account's concurrent session limit.

This requirement is met through the pam_limits(8) module.

Both new pam modules require deeper intergration between middlewared
sessions and the linux PAM framework and so this commit extensively
reworks how sessions are managed in TrueNAS. The scope of this work is
documented within the middlewared developer documentation within the
authentication section.