NAS-131474 / 25.10 / Upload multiple files (#16368)
The lxc integration has exposed a feature that we do not have with job
framework. We need to be able to upload 2 different files at once using
the upload endpoint.

The use-case is simple, distrobuilder will create a metadata file and a
rootfs image. Both of these files are needed to “import” and
subsequently run the container. This will be important for the VEEAM
container image that is being worked on by solutions team.

We need to change the middleware job framework so that it accepts >= 2
files.