NAS-135715 / 25.10 / default accept_ra sysctl to 2 (#16442)
Commit 3b9b1c76032a3f50053fce3c8255f178e4ba3c34 subtly broke users who
depend on SLAAC related functionality. When ipv6 was enabled, it causes
the net.ipv6.conf.*.forwarding sysctl to get set to 1 on _ALL_
interfaces on the host. This causes the kernel to ignore any router
advertisements on those interfaces because it treats the host as a
"router". To fix this problem we need to set `accept_ra=2` which will
tell the kernel to accept RA's on interfaces when forwarding=1.