NAS-135715 / 25.04.2 / default accept_ra sysctl to 2 (by yocalebo) (#16443)
Commit 3b9b1c76032a3f50053fce3c8255f178e4ba3c34 subtly broke users who
depend on SLAAC related functionality. When ipv6 was enabled, it causes
the net.ipv6.conf.*.forwarding sysctl to get set to 1 on _ALL_
interfaces on the host. This causes the kernel to ignore any router
advertisements on those interfaces because it treats the host as a
"router". To fix this problem we need to set `accept_ra=2` which will
tell the kernel to accept RA's on interfaces when forwarding=1.

Original PR: https://github.com/truenas/middleware/pull/16442
Jira URL: https://ixsystems.atlassian.net/browse/NAS-135715

Co-authored-by: Caleb <yocalebo@gmail.com>