NAS-135772 / 25.10 / Make sure missing packages are pulled in build as incus needs it with goldeye mirrors (#16457)
## Problem

VMs on incus need multiple packages which are missing right now after
the switch to goldeye mirrors.

## Solution

Make sure ovmf/qemu-utils are pulled in as VMs on incus won't work
otherwise. We also need qemu-system-modules-spice package as it seems
there is no global way to disable spice for VM instances and we should
probably investigate how best to do that (currently it runs on a unix
socket).