NAS-135664 / 25.10 / Persist periodic snapshot task state (#16459)
Additionally, persist replication task state on every state change
(instead of doing it every hour).

This was introduced ages ago when we tried to protect the USB root
devices from writing too frequently in edge cases (replication tasks ran
every 5 minutes or things like that). Now this does more bad than good
and results in task state not being persisted if the system is rebooted
abruptly.

The changes are too big to introduce them to 25.04.