NAS-135643 / 25.10 / Fix API calls when connected to /websocket (#16444)
Connecting to /websocket will attempt to use the 24.10 version of the
API.

For all old versions of the API, we use factory functions to lazily load
the version when it is needed. When a method model is not found in an
older version (e.g. 24.10), we register it to the older version with a
value of `None` since `None` is returned by this function when a
`KeyError` is raised.

https://github.com/truenas/middleware/blob/accfe25027d247ab4e0cac2ba42a34eb6ca41972/src/middlewared/middlewared/main.py#L1385
The solution is to instead allow `KeyError` to propagate so that we make
use of the `passthrough_nonexistent_methods` feature and use the most
recent version of the method.

While here, I've also resolved a related FIXME comment and added more
type hints.

http://jenkins.eng.ixsystems.net:8080/job/tests/job/api_tests/4335/