NAS-135792 / 25.04.2 / Fix mutual TLS authentication for LDAP (#16466)
This commit fixes a longstanding validation bug in mutual tls
authentication for the LDAP plugin. Tests will be added once CI pipeline
added with this authentication type and an LDAP identity provider. This
commit is not targeted for master since the underlying fix will be
rolled into the overall larger rewrite project.