NAS-135898 / 25.10 / Add special case handling of user password disable. (#16508)
When user passwords are disabled the user can no longer manage password
change requirements or any password policy restrictions. Under certain
configurations, e.g. STIG, this can lead to the user account being
disabled. We want to avoid this condition.

The fix for this is to special case the user shadow entry to disable all
expiration type settings when the user has password authentication
disabled.

The shadow entry for a password disabled user will be `*:::::::`. This
will prevent the user getting disabled due to password requirements.

Also added a CI unit test.
Also fixed a small error in the unit test. Reporting info on an assert
was missing some data.

This fixes the issue raised in
[NAS-135863](https://ixsystems.atlassian.net/browse/NAS-135863)