NAS-135952 / 25.10 / Fix password expiry check (#16526)
The PAM subsystem properly detects and processes an expired password
condition. By default, the condition is reported as a general failure.
The details of the error are included in failure messages. We want to
properly present the expiry condition to the user. The password expiry
check is performed in the 'account' section of PAM processing.
We were not detecting the expiry condition because we were looking for
it in the PAM 'authentication' section.

**The fix:**
Moved the expiry trap to after the pam account processing call.

**Also updated the CI unit tests:**
* Ensure clean state going in and exiting.
* Small tweak to `test__password_expiry_warning` to avoid false assert.

CI unit test
[here](http://jenkins.eng.ixsystems.net:8080/job/tests/job/unit_tests/1396/)