NAS-135946 / 25.10 / Allow password aging bypass for admin OTPW (#16529)
This commit fixes a usability issue with generating temporary passwords
for user accounts while in STIG compatibility mode with a minimum
password age set. There are various situations in which an account
administrator may need to generate a password for users so that they can
authenticate and then set their password and two-factor settings. When
the OTPW has been generated in this manner, set a flag that is then used
to bypass the minimum password age validation.