NAS-135991 / 25.10 / optimize pool.dataset.path_in_locked_datasets (#16524)
This highly optimizes the call to
`pool.dataset.path_in_locked_datasets`. This is done by utilizing the
newly minted `truenas_pylibzfs` module.

This is taking a completely different approach for how we determine if a
path is in a locked resource. Instead of enumerating all encrypted zfs
filesystems/volumes on the system and then iterating over that list and
see if the path is a child of any of the datasets, we will simply open
each component in the `path` to see if it's encrypted and locked.

Because the new `truenas_pylibzfs` module is thread-safe, we're able to
do this in our thread-pool. This is a substantive improvement in
efficiency and speed compared to what we were doing previously.

---------

Co-authored-by: Logan Cary <logan.cary@ixsystems.com>