NAS-135929 / 25.10 / Virt volume id should be unique (#16515)
## Problem

We are able to do the following in incus:
```
❯ incus storage volume create tank2 cli1 --type=block size=1GiB
Storage volume cli1 created
❯ incus storage volume create tank cli1 --type=block size=1GiB
Storage volume cli1 created
```

This means that the same name can be shared across multiple storage
pools for a volume and it will break how we allow users to add a volume
to an instance.

## Solution

Make sure that consumers are correctly able to consume a pool with an
instance when we have a volume of the same name across multiple storage
pools so the proposed idea is to make the `id` which we use in
middleware unique by adding the storage pool to it.