NAS-135501 / 25.10 / Gracefully handle misconfigured incus instances (#16509)
## Problem

If any attachment of a virt instance is somehow deleted/removed and
incus is started again, incus refuses to initialize and we get in an
erroneous start. What happens behind the scenes is that incus iterates
over it's storage pools and goes over it's containers/vms and then tries
to load it's config and assert attachment's existence and if it is not
there, it errors out.

## Solution

What we do is that we identify these instances from the error incus
gives us and then move it to another dataset and log it that it has been
moved because incus was not able to initialize it. Ideally incus should
still cater to other instances and mark that instance in a
failed/erroneous state but until that happens upstream, we handle that
gracefully on our side.