NAS-135667 / 25.04.2 / Fix update.get_pending (by yocalebo) (#16557)
UI team has switched to using this endpoint which has exposed a bug that
has probably been here since it was written. For reasons unknown,
systems running version can be the same as the file that has been
downloaded. In this instance, this returned an "update" but the versions
were the same. Regardless of that the `can_update` function has been
written for this exact scenario.

Original PR: https://github.com/truenas/middleware/pull/16409
Jira URL: https://ixsystems.atlassian.net/browse/NAS-135667

---------

Co-authored-by: Caleb <yocalebo@gmail.com>