Improve alert text for SMB users missing passwords
Our webui does not display the alert title which renders the original
alert for this mishap ambiguous to the end-user when viewed through
the webui. This commit reworks the alert text to simplify the
sentence syntax and clarify the nature of the issue.

(cherry picked from commit 6bde71d0c04be9a68596c4e045e670b27f4f9bb4)