NAS-136070 / 25.04.2 / Improve alert text for SMB users missing passwords (by anodos325) (#16565)
Our webui does not display the alert title which renders the original
alert for this mishap ambiguous to the end-user when viewed through the
webui. This commit reworks the alert text to simplify the sentence
syntax and clarify the nature of the issue.

Original PR: https://github.com/truenas/middleware/pull/16561
Jira URL: https://ixsystems.atlassian.net/browse/NAS-136070

Co-authored-by: Andrew Walker <awalker@ixsystems.com>