NAS-136065 / 25.10 / Fix pam_faillock ordering (#16563)
This comit fixes several issues with pam_faillock configuration in the
common-auth pam include file.

When pam_faillock is configured in common-auth and pam_oath is
configured in the SSH pam configuration file, the pam_faillock
sufficient control causes authentication to short-circuit potentially
bypassing oath authentication.

This commit changes common-auth-unix file so that it is dynamically
generated depending on whether pam_faillock is enabled. If pam_faillock
is enabled, then the the pam_deny line is replaced with the pam_faillock
failure line.

This commit changes the control directive for the pam_faillock success
line so that it is no longer sufficient (avoiding the short-circuit).