NAS-136003 / 25.10 / Refactor directory services related plugins to new API (#16552)
This commit significantly redesigns how directory services work on
TrueNAS while also converting the related plugins to the new API schema.

* The activedirectory and ldap plugins are removed in favor of a single
directoryservices ConfigService plugin. A thin wrapper for
activedirectory.config/update and ldap.config/update is preserved to
avoid breaking the webui entirely until the webui can be updated to use
the new interfaces.

* The idmap CRUD plugin is removed and replaced with an array of trusted
domains in the activedirectory configuration in the directoryservices
plugin. A small shim is left in place to provide fake idmap.query
results in order to avoid breaking the webui.

* Directoryservices.get_state is deprecated in favor of
directoryservices.status.

* Tests related to directory services are moved to
`tests/directory_services` and runtest.py is updated to allow for
running tests that are under directories that aren't tests/api2. This is
preparatory step for adding explicit failover tests.