NAS-136123 / 25.10 / Use new libzfs module for quotas (#16574)
This replaces the `zfs.dataset.get/set_quotas` and
`query_for_quota_alert` private endpoints in the `zfs.dataset`
namespace.

The reason why it was done this way was to be 100% backwards compatible
with API so the UI team did not have to make any changes.

On an under-powered VM with 1 dataset, I'm seeing about a 3-4x speed
improvement for getting/setting quotas and almost 2x for querying
datasets for the quota alert. The benefits of this will be even more
pronounced on a real system with real-world data.