NAS-136149 / 25.10 / Use default for NFS RPC pipefs-directory (#16580)
The default path for NFS rpc idmapd is `/var/lib/nfs/rpc_pipefs`. We are
overriding this in `nfs.conf` with the legacy setting `/run/rpc_pipefs`.

Using the legacy path with our startup process and Linux 6.12 has
exposed a deficiency in using the transitory paths in the `/run`
directory where the `idmap` daemon cannot start because the legacy path
is not available.

The fix:
Use the default setting for `pipefs-directory`.

Tested manually.